+++ ios/Classes/BrazilianLocationsPlugin.swift
@@ -1,0 +1,19 @@
+import Flutter
+import UIKit
+
+public class BrazilianLocationsPlugin: NSObject, FlutterPlugin {
+  public static func register(with registrar: FlutterPluginRegistrar) {
+    let channel = FlutterMethodChannel(name: "brazilian_locations", binaryMessenger: registrar.messenger())
+    let instance = BrazilianLocationsPlugin()
+    registrar.addMethodCallDelegate(instance, channel: channel)
+  }
+
+  public func handle(_ call: FlutterMethodCall, result: @escaping FlutterResult) {
+    switch call.method {
+    case "getPlatformVersion":
+      result("iOS " + UIDevice.current.systemVersion)
+    default:
+      result(FlutterMethodNotImplemented)
+    }
+  }
+}