+++ CHANGELOG.md
@@ -1,12 +1,19 @@
-## [3.0.0] - 1 Mar 2024
+## [3.1.0] - 8 Mar 2024
+- Option to customize style for text field.
+- Option to use label above the dropdown component.
+- Plugins added for all platforms.
+- Simplified JSON file, reducing the size of the package.
+- Fixed conflict dependency with the HTTP package.
+
+## [3.0.0] - 7 Mar 2024
 - Added option to load JSON from IBGE API.
 - Added support for all platforms.
 - Fixed a bug related to label.
 
-## [2.0.0] - 1 Mar 2024
+## [2.0.0] - 4 Mar 2024
 - Build a stable Release
 
-## [1.1.0] - 1 Mar 2024
+## [1.1.0] - 4 Mar 2024
 - Fixing some bugs
 - Update README file